+++ src/test/java/org/bigml/binding/DatasetsStepdefs.java
@@ -74,7 +74,6 @@ public class DatasetsStepdefs {
     }
 
 
-
     @Then("^I create a dataset from the cluster and the centroid$")
     public void I_create_a_dataset_from_the_cluster_and_the_centroid() throws Throwable {
         String clusterId = (String) context.cluster.get("resource");
@@ -91,7 +90,6 @@ public class DatasetsStepdefs {
         context.dataset = (JSONObject) resource.get("object");
         commonSteps.the_resource_has_been_created_with_status(context.status);
     }
-
 
     @Given("^I wait until the dataset status code is either (\\d) or (\\d) less than (\\d+)")
     public void I_wait_until_dataset_status_code_is(int code1, int code2,
@@ -318,5 +316,38 @@ public class DatasetsStepdefs {
         }
 
     }
+    
+    
+    @Then("^I create a dataset associated to centroid \"(.*)\"$")
+    public void I_create_a_dataset_associated_to_centroid(String centroidId) 
+    		throws Throwable {
+        String clusterId = (String) context.cluster.get("resource");
+        
+        JSONObject args = new JSONObject();
+        args.put("tags", Arrays.asList("unitTest"));
+        args.put("centroid", centroidId);
 
+        JSONObject resource = BigMLClient.getInstance().createDataset(
+        		clusterId, args, 5, null);
+        context.status = (Integer) resource.get("code");
+        context.location = (String) resource.get("location");
+        context.dataset = (JSONObject) resource.get("object");
+        commonSteps.the_resource_has_been_created_with_status(context.status);
+    }
+    
+    @Then("^the dataset is associated to the centroid \"([^\"]*)\" of the cluster$")
+    public void the_dataset_is_associated_to_the_centroid_of_the_cluster(String centroid) 
+    		throws Throwable {
+        
+    	JSONObject resource = BigMLClient.getInstance().getCluster(
+            (String) context.cluster.get("resource"));
+
+        context.status = (Integer) resource.get("code");
+        assertEquals(AbstractResource.HTTP_OK, context.status);
+        
+        assertEquals(context.dataset.get("resource"), 
+        	String.format("dataset/%s",
+                Utils.getJSONObject(resource,
+                        String.format("object.cluster_datasets.%s", centroid) )));
+    }
 }